pyproject.toml
[tool.poetry]
name = "midi-nn"
version = "0.1.0"
description = ""
authors = ["paul <dev@studiop.page>"]

[tool.poetry.dependencies]
python = "^3.8"
note-seq = "^0.0.1"
tensorflow-gpu = "^2.3.1"
jupyterlab = "^2.2.8"
notebook = "^6.1.4"
jupyter_http_over_ws = "^0.0.8"
keras = "^2.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
